Abstract out logic in autocomplete ui to a manager class
