+++ src/main/java/cs/f10/t1/nursetraverse/ui/MainWindow.java
@@ -41,6 +41,7 @@ public class MainWindow extends UiPart<Stage> {
     private CommandBox commandBox;
     private DataPanelsTabPaneManager dataPanelsTabPaneManager;
     private HistoryPanel historyPanel;
+    private AutoCompletePanelManager autoCompletePanelManager;
 
     @FXML
     private StackPane commandBoxPlaceholder;
@@ -142,21 +143,22 @@ public class MainWindow extends UiPart<Stage> {
         StatusBarFooter statusBarFooter = new StatusBarFooter(logic.getPatientBookFilePath());
         statusbarPlaceholder.getChildren().add(statusBarFooter.getRoot());
 
-        autoCompletePanel = new AutoCompletePanel(
+        autoCompletePanel = new AutoCompletePanel();
+        autoCompletePanelPlaceholder.getChildren().add(autoCompletePanel.getRoot());
+        autoCompletePanelManager = new AutoCompletePanelManager(
+                autoCompletePanel,
                 logic.getFilteredPatientList(),
                 logic.getFilteredAppointmentList(),
-                logic.getObservableHistoryList()
-        );
-        autoCompletePanelPlaceholder.getChildren().add(autoCompletePanel.getRoot());
+                logic.getObservableHistoryList());
 
         commandBox = new CommandBox(this::executeCommand);
         commandBoxPlaceholder.getChildren().add(commandBox.getRoot());
 
         // Add observers into commandBox observer list
-        commandBox.addObserver(autoCompletePanel);
+        commandBox.addObserver(autoCompletePanelManager);
         commandBox.addObserver(resultDisplay);
         // Set data sender for commandBox
-        commandBox.setDataSender(autoCompletePanel);
+        commandBox.setDataSender(autoCompletePanelManager);
 
         dataPanelsTabPaneManager = new DataPanelsTabPaneManager(dataPanelsTabPane,
                 patientTabPage,
@@ -252,12 +254,5 @@ public class MainWindow extends UiPart<Stage> {
             resultDisplay.setFeedbackToUser(e.getMessage());
             throw e;
         }
-    }
-
-    /**
-     * Set listeners for all individual components in main window
-     */
-    public void setAllListeners() {
-        commandBox.setOnButtonPressedListener();
     }
 }
