+++ src/main/java/cs/f10/t1/nursetraverse/ui/UiManager.java
@@ -41,8 +41,6 @@ public class UiManager implements Ui {
             mainWindow = new MainWindow(primaryStage, logic);
             mainWindow.show(); //This should be called before creating other UI parts
             mainWindow.fillInnerParts();
-            // Main window listener
-            mainWindow.setAllListeners();
 
         } catch (Throwable e) {
             logger.severe(StringUtil.getDetails(e));
