+++ src/main/java/cs/f10/t1/nursetraverse/ui/UiObserver.java
@@ -6,7 +6,7 @@ import javafx.scene.input.KeyCode;
  * An observer interface to observe the command box class for userinputs
  * This interface was created so as to provide a facade as ui interacts with one another
  */
-public interface ObserverUi {
+public interface UiObserver {
     /**
      * Updates observer after key press activity
      * @param keyCode key code that was pressed by user
